Refactor utility functions for front end and set today date for plan template page
