+++ assets/js/search.js
@@ -1,5 +1,9 @@
 // methods for the search page
-import { logOut, updateUsername } from "./user.js";
+import { locateMe, setDateToday } from "./utils.js";
+
+import { logOut } from "./user.js";
+
+setDateToday();
 
 (function ($) {
     $("#location").autocomplete(
@@ -9,7 +13,7 @@ import { logOut, updateUsername } from "./user.js";
                     {
                         url: "/v1/cities",
                         dataType: "json",
-                        data: {term: request.term},
+                        data: { term: request.term },
                         success: function (data) {
                             response($.map(data.results, function (city) {
                                 if (city.region) {
@@ -29,56 +33,6 @@ import { logOut, updateUsername } from "./user.js";
 document.getElementById("logout-confirm-btn").addEventListener(
     "click", logOut
 )
-
-function locateMe() {
-    async function success(location) {
-        const latitude = location.coords.latitude;
-        const longitude = location.coords.longitude;
-
-        console.log(`latitude ${latitude} and longitude: ${longitude}`);
-
-        const url = "/v1/reverse-geocoding"
-        await axios.get(url, {
-            params: {
-                lat: latitude,
-                lng: longitude
-            }
-        })
-            .then(
-                response => {
-                    const reverseGeocodingResults = response.data.results;
-                    document.getElementById("location").value = [reverseGeocodingResults.city, reverseGeocodingResults.admin_area_level_one, reverseGeocodingResults.country].join(", ");
-                }
-            ).catch(
-                err => console.error(err)
-            )
-    }
-
-    function error() {
-    }
-
-    if (navigator.geolocation) {
-        navigator.geolocation.getCurrentPosition(success, error);
-    }
-}
-
-function setDateToday() {
-    const today = new Date();
-    console.log("today's date is: " + today);
-    let month = today.getMonth() + 1;
-    if (month < 10) {
-        month = "0" + month.toString();
-    }
-    let day = today.getDate();
-    if (day < 10) {
-        day = "0" + day.toString();
-    }
-    document.getElementById("datepicker").value = [today.getFullYear(), month, day].join("-");
-}
-
-const username = updateUsername();
-
-setDateToday();
 
 document.querySelector('#autofill').addEventListener('click', locateMe);
