+++ assets/js/user.js
@@ -45,4 +45,7 @@ function logOut() {
     location.reload();
 }
 
+const username = updateUsername();
+document.getElementById("profile").addEventListener("click", () => window.location = `/v1/users/${username}/profile`);
+
 export { updateUsername, logOut }
