+++ templates/plan_template.html
@@ -137,7 +137,8 @@
         </div>
 
 
-        <table class="table table-sm table-bordered d-none" style="background: lightcyan; table-layout: fixed" id="results">
+        <table class="table table-sm table-bordered d-none" style="background: lightcyan; table-layout: fixed"
+            id="results">
             <thead class="table-light">
                 <tr>
                     <th>Time</th>
@@ -160,6 +161,7 @@
 
     <script type="module" src="assets/js/jwt-decode.js"></script>
     <script type="module" src="assets/js/user.js"></script>
+    <script type="module" src="assets/js/utils.js"></script>
 </body>
 
 </html>
