+++ templates/search_page.html
@@ -155,6 +155,7 @@
     <script src="https://cdn.jsdelivr.net/npm/js-cookie@rc/dist/js.cookie.min.js"></script>
     <script type="module" src="assets/js/jwt-decode.js"></script>
     <script type="module" src="assets/js/user.js"></script>
+    <script type="module" src="assets/js/utils.js"></script>
     <script type="module" src="assets/js/search.js"></script>
 
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.6.0/jquery.min.js"></script>
